+++ src/Fabricator.Core/Environment/SystemPlatform.cs
@@ -5,4 +5,8 @@ namespace Fabricator.Core.Environment;
 public sealed class SystemPlatform : ISystemPlatform
 {
     public bool IsMacOS => RuntimeInformation.IsOSPlatform(OSPlatform.OSX);
+
+    public bool IsWindows => RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
+
+    public bool IsLinux => RuntimeInformation.IsOSPlatform(OSPlatform.Linux);
 }